For rx admin

+++ src/directives/post-relation-cascade.ts
@@ -26,9 +26,11 @@ export class PostRelationCascadeDirective extends PostDirective {
       [relationMetaCollection.relationName]: {},
       '@getOne': true,
     });
-    this.oldRelationIds = data.data[relationMetaCollection.relationName]?.map(
-      (relation: { id: number }) => relation.id,
-    );
+    if (data.data) {
+      this.oldRelationIds = data.data[relationMetaCollection.relationName]?.map(
+        (relation: { id: number }) => relation.id,
+      );
+    }
   }
 
   //该命令的权限通过magicService完成
@@ -40,8 +42,11 @@ export class PostRelationCascadeDirective extends PostDirective {
     const deleteIds = this.oldRelationIds.filter(
       (id) => !savedInstances.find((instance) => instance.id === id),
     );
-    await this.magicService.delete({
-      [relationMetaCollection.entity]: deleteIds,
-    });
+
+    if (deleteIds.length > 0) {
+      await this.magicService.delete({
+        [relationMetaCollection.entity]: deleteIds,
+      });
+    }
   }
 }
